Commit title : Reorganizing files
What : Reorganizing main nav bar files and renaming some route in Routes.dart

Why : for ความสวยงาม

How : create folder for main route and remove "screen" from routes name
